ingestion/requirements.txt
pandas>=2.2
pyarrow>=16.0
requests>=2.32
tenacity>=8.2
python-dateutil>=2.9
google-cloud-storage>=2.17
